+++ exporters/trace/zipkin/model_test.go
@@ -60,7 +60,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.String("attr2", "bar"),
 				attribute.Array("attr3", []int{0, 1, 2}),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -93,7 +93,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -129,7 +129,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -165,7 +165,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -204,7 +204,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.String("net.peer.ip", "1.2.3.4"),
 				attribute.Int64("net.peer.port", 9876),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -240,7 +240,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -276,7 +276,7 @@ func TestModelConversion(t *testing.T) {
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",
@@ -334,7 +334,7 @@ func TestModelConversion(t *testing.T) {
 			Attributes: []attribute.KeyValue{
 				attribute.String("error", "false"),
 			},
-			MessageEvents: []trace.Event{
+			MessageEvents: []tracesdk.Event{
 				{
 					Time: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
 					Name: "ev1",